Started Statistics Fragment
This commit includes the creation of the fragment for the statistics. It currently includes a tab layout and a viewpager for all the information. Take note this is only a skeleton and has no functionality. Alongside this, the charts have not been added yet, as I am still learning how to make them.