Implementation of History Fragment and Continuation of Statistics Fragment
I apologize for the large commit as I forgot to commit multiple times and did all of these changes at once. This commit includes several changes over different areas.

For the statistics fragment, the tabLayout has been colored appropriately, and a layout for the frequency chart has been implemented. Note that the charts haven't been implemented themselves, but are the last objects that have to be finished.

For the history fragment, it has been implemented and completed. It is very bare-bones, so recommendations on how to stylize it will be heavily appreciated.

On other changes, the settings has been temporarily removed, as we have not declared what we would need to add in the settings, if at all. Please discuss this via a meet in a future time, but it will not be part of the MVP until further notice.

The commit after this will (hopefully) conclude the development of the android studio application, and I will move on with developing the backend and database.